Refactor to RedirectRouteCollectionLoader
