+++ SuluRedirectBundle.php
@@ -12,7 +12,6 @@
 namespace Sulu\Bundle\RedirectBundle;
 
 use Sulu\Bundle\PersistenceBundle\PersistenceBundleTrait;
-use Sulu\Bundle\RedirectBundle\DependencyInjection\RouteDefaultOptionsCompilerPass;
 use Sulu\Bundle\RedirectBundle\Entity\RedirectRoute;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
@@ -26,10 +25,6 @@ class SuluRedirectBundle extends Bundle
 
     public function build(ContainerBuilder $container): void
     {
-        $container->addCompilerPass(
-            new RouteDefaultOptionsCompilerPass('sulu_redirect.routing.provider', 1)
-        );
-
         $this->buildPersistence(
             [
                 RedirectRoute::class => 'sulu.model.redirect_route.class',
